moving currency-related code to a separate package
