+++ res/layout/activity_currency_chart.xml
@@ -24,7 +24,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.money.manager.ex.CurrencyChartActivity"
+    tools:context="com.money.manager.ex.currency.CurrencyChartActivity"
     android:orientation="horizontal" >
 
     <ImageView
