+++ src/com/money/manager/ex/currency/CurrencyFormatsActivity.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package com.money.manager.ex;
+package com.money.manager.ex.currency;
 
 import android.content.ContentValues;
 import android.content.Intent;
@@ -27,6 +27,8 @@ import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.Toast;
 
+import com.money.manager.ex.Constants;
+import com.money.manager.ex.R;
 import com.money.manager.ex.database.TableCurrencyFormats;
 import com.money.manager.ex.fragment.BaseFragmentActivity;
 import com.money.manager.ex.utils.CurrencyUtils;
